add conditional tasks depending on if environment is dev or production

+++ lib/tasks/dev.rake
@@ -2,12 +2,17 @@ desc "Fill the database tables with some sample data"
 task({ :sample_data => :environment }) do
   puts "Sample data task running"
   
-  ActiveRecord::Base.connection.tables.each do |t|
-    ActiveRecord::Base.connection.reset_pk_sequence!(t)
+  if Rails.environment.dev?
+    User.destroy_all
+    Board.destroy_all
+    Post.destroy_all
   end
-
-  Board.destroy_all
-  Post.destroy_all
+  
+  if Rails.environment.production?
+    ActiveRecord::Base.connection.tables.each do |t|
+      ActiveRecord::Base.connection.reset_pk_sequence!(t)
+    end
+  end
   
   usernames= ["alice", "bob", "carol", "dave", "eve"]
